# docs/api-docs.yaml
openapi: 3.0.1
info:
  title: cambusa-controller
  description: Servizio per la tracciatura delle scadenze dei prodotti
  contact:
    name: /stmsat/cambusa-backend
    url: https://github.com/stmsat/cambusa-backend
servers:
- url: http://127.0.0.1:8080
  description: Generated server url
paths:
  /cambusa/tipi:
    get:
      tags:
      - cambusa-controller
      description: "Lista dei tipi, ricercabile per nome e ordinabile"
      operationId: getTipi
      parameters:
      - name: name
        in: query
        required: false
        schema:
          type: string
      - name: sortby
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
          default:
          - name
      - name: sortdirection
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
          default:
          - ASC
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tipo'
    put:
      tags:
      - cambusa-controller
      description: Salva un TIpo (insert o update)
      operationId: putTipo
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tipo'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
  /cambusa/prodotti:
    get:
      tags:
      - cambusa-controller
      description: "Lista dei prodotti, ricercabile per vari parametri"
      operationId: getProdotti
      parameters:
      - name: name
        in: query
        required: false
        schema:
          type: string
          default: ""
      - name: dataScadenzaLt
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: dataScadenzaGenerataLt
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: dataScadenzaGt
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: dataScadenzaGenerataGt
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: tipo
        in: query
        required: false
        schema:
          type: array
          items:
            $ref: '#/components/schemas/Tipo'
      - name: posizione
        in: query
        required: false
        schema:
          type: array
          items:
            $ref: '#/components/schemas/Posizione'
      - name: sortby
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
          default:
          - name
      - name: sortdirection
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
          default:
          - ASC
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prodotto'
    put:
      tags:
      - cambusa-controller
      description: Salva un Prodotto (insert o update)
      operationId: putProdotto
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Prodotto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
  /cambusa/posizioni:
    get:
      tags:
      - cambusa-controller
      description: "Lista delle posizioni, ricercabile per nome e ordinabile"
      operationId: getPosizioni
      parameters:
      - name: name
        in: query
        required: false
        schema:
          type: string
      - name: sortby
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
          default:
          - name
      - name: sortdirection
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
          default:
          - ASC
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Posizione'
    put:
      tags:
      - cambusa-controller
      description: Salva una Posizione (insert o update)
      operationId: putPosizione
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Posizione'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
  /cambusa/prodotti/{id}:
    get:
      tags:
      - cambusa-controller
      description: Prodotto per Id
      operationId: getProdotto
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prodotto'
    delete:
      tags:
      - cambusa-controller
      description: Elimina un Prodotto
      operationId: deleteProdotto
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
    patch:
      tags:
      - cambusa-controller
      description: Aggiorna singoli campi di un Prodotto
      operationId: patchProdotto
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: aperto
        in: query
        required: false
        schema:
          type: boolean
      - name: dataApertura
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: quantita
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prodotto'
  /cambusaDevel/init:
    get:
      tags:
      - cambusa-controller-devel
      description: Inizializza il dataset di test
      operationId: init
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
  /cambusa/tipi/{id}:
    get:
      tags:
      - cambusa-controller
      description: Tipo per Id
      operationId: getTipo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tipo'
    delete:
      tags:
      - cambusa-controller
      description: Elimina un Tipo
      operationId: deleteTipo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
  /cambusa/posizioni/{id}:
    get:
      tags:
      - cambusa-controller
      description: Posizione per Id
      operationId: getPosizione
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Posizione'
    delete:
      tags:
      - cambusa-controller
      description: Elimina una Posizione
      operationId: deletePosizione
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
    Tipo:
      required:
      - dataScadenzaPreferibile
      - name
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        dataScadenzaPreferibile:
          type: boolean
        giorniValiditaDopoScadenza:
          type: integer
          format: int32
        apribile:
          type: boolean
        giorniValiditaDopoApertura:
          type: integer
          format: int32
    Posizione:
      required:
      - name
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    Prodotto:
      required:
      - dataScadenza
      - name
      - tipo
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        dataScadenza:
          type: string
          format: date
        tipo:
          $ref: '#/components/schemas/Tipo'
        posizione:
          $ref: '#/components/schemas/Posizione'
        dataApertura:
          type: string
          format: date
        aperto:
          type: boolean
        quantita:
          minimum: 1
          type: integer
          format: int32
        dataScadenzaGenerata:
          type: string
          format: date